requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
boto3==1.10.45
botocore==1.13.45
celery==4.4.1
choicesenum==0.6.0
dj-database-url==0.5.0
Django==3.0.1
django-exiffield==2.1.0
django-storages==1.8
docutils==0.15.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
future==0.18.2
gunicorn==20.0.4
jmespath==0.9.4
joblib==0.14.1
jsonfield2==3.1.0
kombu==4.6.8
numpy==1.18.1
Pillow==6.2.1
playsound==1.2.2
psycopg2-binary==2.8.4
pydub==0.23.1
pyglet==1.5.0
pyobjc==6.1
pyobjc-core==6.1
pyobjc-framework-Accounts==6.1
pyobjc-framework-AddressBook==6.1
pyobjc-framework-AdSupport==6.1
pyobjc-framework-AppleScriptKit==6.1
pyobjc-framework-AppleScriptObjC==6.1
pyobjc-framework-ApplicationServices==6.1
pyobjc-framework-AuthenticationServices==6.1
pyobjc-framework-Automator==6.1
pyobjc-framework-AVFoundation==6.1
pyobjc-framework-AVKit==6.1
pyobjc-framework-BusinessChat==6.1
pyobjc-framework-CalendarStore==6.1
pyobjc-framework-CFNetwork==6.1
pyobjc-framework-CloudKit==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-Collaboration==6.1
pyobjc-framework-ColorSync==6.1
pyobjc-framework-Contacts==6.1
pyobjc-framework-ContactsUI==6.1
pyobjc-framework-CoreAudio==6.1
pyobjc-framework-CoreAudioKit==6.1
pyobjc-framework-CoreBluetooth==6.1
pyobjc-framework-CoreData==6.1
pyobjc-framework-CoreHaptics==6.1
pyobjc-framework-CoreLocation==6.1
pyobjc-framework-CoreMedia==6.1
pyobjc-framework-CoreMediaIO==6.1
pyobjc-framework-CoreML==6.1
pyobjc-framework-CoreMotion==6.1
pyobjc-framework-CoreServices==6.1
pyobjc-framework-CoreSpotlight==6.1
pyobjc-framework-CoreText==6.1
pyobjc-framework-CoreWLAN==6.1
pyobjc-framework-CryptoTokenKit==6.1
pyobjc-framework-DeviceCheck==6.1
pyobjc-framework-DictionaryServices==6.1
pyobjc-framework-DiscRecording==6.1
pyobjc-framework-DiscRecordingUI==6.1
pyobjc-framework-DiskArbitration==6.1
pyobjc-framework-DVDPlayback==6.1
pyobjc-framework-EventKit==6.1
pyobjc-framework-ExceptionHandling==6.1
pyobjc-framework-ExecutionPolicy==6.1
pyobjc-framework-ExternalAccessory==6.1
pyobjc-framework-FileProvider==6.1
pyobjc-framework-FileProviderUI==6.1
pyobjc-framework-FinderSync==6.1
pyobjc-framework-FSEvents==6.1
pyobjc-framework-GameCenter==6.1
pyobjc-framework-GameController==6.1
pyobjc-framework-GameKit==6.1
pyobjc-framework-GameplayKit==6.1
pyobjc-framework-ImageCaptureCore==6.1
pyobjc-framework-IMServicePlugIn==6.1
pyobjc-framework-InputMethodKit==6.1
pyobjc-framework-InstallerPlugins==6.1
pyobjc-framework-InstantMessage==6.1
pyobjc-framework-Intents==6.1
pyobjc-framework-IOSurface==6.1
pyobjc-framework-iTunesLibrary==6.1
pyobjc-framework-LatentSemanticMapping==6.1
pyobjc-framework-LaunchServices==6.1
pyobjc-framework-libdispatch==6.1
pyobjc-framework-LinkPresentation==6.1
pyobjc-framework-LocalAuthentication==6.1
pyobjc-framework-MapKit==6.1
pyobjc-framework-MediaAccessibility==6.1
pyobjc-framework-MediaLibrary==6.1
pyobjc-framework-MediaPlayer==6.1
pyobjc-framework-MediaToolbox==6.1
pyobjc-framework-MetalKit==6.1
pyobjc-framework-ModelIO==6.1
pyobjc-framework-MultipeerConnectivity==6.1
pyobjc-framework-NaturalLanguage==6.1
pyobjc-framework-NetFS==6.1
pyobjc-framework-Network==6.1
pyobjc-framework-NetworkExtension==6.1
pyobjc-framework-NotificationCenter==6.1
pyobjc-framework-OpenDirectory==6.1
pyobjc-framework-OSAKit==6.1
pyobjc-framework-OSLog==6.1
pyobjc-framework-PencilKit==6.1
pyobjc-framework-Photos==6.1
pyobjc-framework-PhotosUI==6.1
pyobjc-framework-PreferencePanes==6.1
pyobjc-framework-PubSub==6.1
pyobjc-framework-PushKit==6.1
pyobjc-framework-Quartz==6.1
pyobjc-framework-QuickLookThumbnailing==6.1
pyobjc-framework-SafariServices==6.1
pyobjc-framework-SceneKit==6.1
pyobjc-framework-ScreenSaver==6.1
pyobjc-framework-ScriptingBridge==6.1
pyobjc-framework-SearchKit==6.1
pyobjc-framework-Security==6.1
pyobjc-framework-SecurityFoundation==6.1
pyobjc-framework-SecurityInterface==6.1
pyobjc-framework-ServiceManagement==6.1
pyobjc-framework-Social==6.1
pyobjc-framework-SoundAnalysis==6.1
pyobjc-framework-Speech==6.1
pyobjc-framework-SpriteKit==6.1
pyobjc-framework-StoreKit==6.1
pyobjc-framework-SyncServices==6.1
pyobjc-framework-SystemConfiguration==6.1
pyobjc-framework-SystemExtensions==6.1
pyobjc-framework-UserNotifications==6.1
pyobjc-framework-VideoSubscriberAccount==6.1
pyobjc-framework-VideoToolbox==6.1
pyobjc-framework-Vision==6.1
pyobjc-framework-WebKit==6.1
python-dateutil==2.8.1
python-vlc==3.0.7110
pytz==2019.3
s3transfer==0.2.1
scikit-learn==0.22.2.post1
scipy==1.4.1
simpleaudio==1.0.4
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
whitenoise==5.0.1
